# internal/api/server.gen.go
// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen version v1.13.0 DO NOT EDIT.
package api

import (
	"fmt"
	"net/http"

	"github.com/deepmap/oapi-codegen/pkg/runtime"
	"github.com/labstack/echo/v4"
)

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// List all computers
	// (GET /v1/computers)
	GetComputers(ctx echo.Context) error
	// Create data source
	// (POST /v1/computers)
	CreateComputer(ctx echo.Context) error
	// Delete computer
	// (DELETE /v1/computers/{computerName})
	DeleteComputer(ctx echo.Context, computerName string) error
	// Update computer data
	// (PATCH /v1/computers/{computerName})
	UpdateComputer(ctx echo.Context, computerName string) error
}

// ServerInterfaceWrapper converts echo contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler ServerInterface
}

// GetComputers converts echo context to params.
func (w *ServerInterfaceWrapper) GetComputers(ctx echo.Context) error {
	var err error

	// Invoke the callback with all the unmarshalled arguments
	err = w.Handler.GetComputers(ctx)
	return err
}

// CreateComputer converts echo context to params.
func (w *ServerInterfaceWrapper) CreateComputer(ctx echo.Context) error {
	var err error

	// Invoke the callback with all the unmarshalled arguments
	err = w.Handler.CreateComputer(ctx)
	return err
}

// DeleteComputer converts echo context to params.
func (w *ServerInterfaceWrapper) DeleteComputer(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "computerName" -------------
	var computerName string

	err = runtime.BindStyledParameterWithLocation("simple", false, "computerName", runtime.ParamLocationPath, ctx.Param("computerName"), &computerName)
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter computerName: %s", err))
	}

	// Invoke the callback with all the unmarshalled arguments
	err = w.Handler.DeleteComputer(ctx, computerName)
	return err
}

// UpdateComputer converts echo context to params.
func (w *ServerInterfaceWrapper) UpdateComputer(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "computerName" -------------
	var computerName string

	err = runtime.BindStyledParameterWithLocation("simple", false, "computerName", runtime.ParamLocationPath, ctx.Param("computerName"), &computerName)
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter computerName: %s", err))
	}

	// Invoke the callback with all the unmarshalled arguments
	err = w.Handler.UpdateComputer(ctx, computerName)
	return err
}

// This is a simple interface which specifies echo.Route addition functions which
// are present on both echo.Echo and echo.Group, since we want to allow using
// either of them for path registration
type EchoRouter interface {
	CONNECT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	DELETE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	GET(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	HEAD(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	OPTIONS(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	PATCH(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	POST(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	PUT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	TRACE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
}

// RegisterHandlers adds each server route to the EchoRouter.
func RegisterHandlers(router EchoRouter, si ServerInterface) {
	RegisterHandlersWithBaseURL(router, si, "")
}

// Registers handlers, and prepends BaseURL to the paths, so that the paths
// can be served under a prefix.
func RegisterHandlersWithBaseURL(router EchoRouter, si ServerInterface, baseURL string) {

	wrapper := ServerInterfaceWrapper{
		Handler: si,
	}

	router.GET(baseURL+"/v1/computers", wrapper.GetComputers)
	router.POST(baseURL+"/v1/computers", wrapper.CreateComputer)
	router.DELETE(baseURL+"/v1/computers/:computerName", wrapper.DeleteComputer)
	router.PATCH(baseURL+"/v1/computers/:computerName", wrapper.UpdateComputer)

}